+++ src/test/java/kr/modusplant/api/signup/normal/controller/NormalSignUpControllerUnitTest.java
@@ -4,12 +4,12 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import kr.modusplant.api.crud.member.common.util.domain.SiteMemberAuthTestUtils;
 import kr.modusplant.api.crud.member.common.util.domain.SiteMemberTermTestUtils;
 import kr.modusplant.api.crud.member.common.util.domain.SiteMemberTestUtils;
-import kr.modusplant.api.crud.member.domain.service.SiteMemberAuthService;
-import kr.modusplant.api.crud.member.domain.service.SiteMemberService;
-import kr.modusplant.api.crud.member.domain.service.SiteMemberTermService;
+import kr.modusplant.api.crud.member.domain.service.supers.SiteMemberAuthService;
+import kr.modusplant.api.crud.member.domain.service.supers.SiteMemberService;
+import kr.modusplant.api.crud.member.domain.service.supers.SiteMemberTermService;
 import kr.modusplant.api.crud.member.enums.AuthProvider;
 import kr.modusplant.api.crud.term.common.util.domain.TermTestUtils;
-import kr.modusplant.api.crud.term.domain.service.TermService;
+import kr.modusplant.api.crud.term.domain.service.supers.TermService;
 import kr.modusplant.api.signup.normal.model.request.NormalSignUpRequest;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
