+++ src/test/java/kr/modusplant/api/signup/social/service/SocialAuthServiceIntegrationTest.java
@@ -3,9 +3,9 @@ package kr.modusplant.api.signup.social.service;
 import kr.modusplant.api.crud.member.domain.model.SiteMember;
 import kr.modusplant.api.crud.member.domain.model.SiteMemberAuth;
 import kr.modusplant.api.crud.member.domain.model.SiteMemberRole;
-import kr.modusplant.api.crud.member.domain.service.SiteMemberAuthService;
-import kr.modusplant.api.crud.member.domain.service.SiteMemberRoleService;
-import kr.modusplant.api.crud.member.domain.service.SiteMemberService;
+import kr.modusplant.api.crud.member.domain.service.supers.SiteMemberAuthService;
+import kr.modusplant.api.crud.member.domain.service.supers.SiteMemberRoleService;
+import kr.modusplant.api.crud.member.domain.service.supers.SiteMemberService;
 import kr.modusplant.api.crud.member.enums.AuthProvider;
 import kr.modusplant.global.enums.Role;
 import org.junit.jupiter.api.DisplayName;
